[API] Adjust unit tests after upmerge
