+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/AdminUserPassword.xml
@@ -36,14 +36,14 @@
             <operation
                 name="admin_create_reset_password"
                 class="ApiPlatform\Metadata\Post"
-                uriTemplate="/admin/reset-password"
+                uriTemplate="/admin/administrators/reset-password"
                 input="Sylius\Bundle\CoreBundle\Message\Admin\Account\RequestResetPasswordEmail"
             >
                 <denormalizationContext>
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:reset_password:create</value>
+                                <value>sylius:admin:reset_password:create</value>
                             </values>
                         </value>
                     </values>
@@ -58,7 +58,7 @@
             <operation
                 name="admin_update_reset_password"
                 class="ApiPlatform\Metadata\Patch"
-                uriTemplate="/admin/reset-password/{resetPasswordToken}"
+                uriTemplate="/admin/administrators/reset-password/{token}"
                 input="Sylius\Bundle\CoreBundle\Message\Admin\Account\ResetPassword"
                 provider="Sylius\Bundle\ApiBundle\StateProvider\Admin\ResetPasswordProvider"
             >
@@ -66,7 +66,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:reset_password:update</value>
+                                <value>sylius:admin:reset_password:update</value>
                             </values>
                         </value>
                     </values>
