+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/CatalogPromotionTranslation.xml
@@ -33,7 +33,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:catalog_promotion:read</value>
+                                <value>sylius:admin:catalog_promotion:show</value>
                             </values>
                         </value>
                     </values>
