+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/Channel.xml
@@ -33,7 +33,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:channel:read</value>
+                                <value>sylius:admin:channel:show</value>
                             </values>
                         </value>
                     </values>
