+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/Command/SendContactRequest.xml
@@ -25,7 +25,7 @@
         status="202"
     >
         <operations>
-            <operation name="shop_send_contact_request" class="ApiPlatform\Metadata\Post" uriTemplate="/contact-requests">
+            <operation name="shop_send_contact_request" class="ApiPlatform\Metadata\Post" uriTemplate="/contact">
                 <validationContext>
                     <values>
                         <value name="groups">
@@ -39,7 +39,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>shop:contact_request:create</value>
+                                <value>sylius:shop:contact_request:create</value>
                             </values>
                         </value>
                     </values>
