+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/Country.xml
@@ -33,7 +33,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:country:read</value>
+                                <value>sylius:admin:country:show</value>
                             </values>
                         </value>
                     </values>
@@ -45,7 +45,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:country:read</value>
+                                <value>sylius:admin:country:index</value>
                             </values>
                         </value>
                     </values>
@@ -57,7 +57,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:country:create</value>
+                                <value>sylius:admin:country:create</value>
                             </values>
                         </value>
                     </values>
@@ -66,7 +66,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:country:read</value>
+                                <value>sylius:admin:country:show</value>
                             </values>
                         </value>
                     </values>
@@ -78,7 +78,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:country:update</value>
+                                <value>sylius:admin:country:update</value>
                             </values>
                         </value>
                     </values>
@@ -87,7 +87,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:country:read</value>
+                                <value>sylius:admin:country:show</value>
                             </values>
                         </value>
                     </values>
@@ -99,7 +99,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>shop:country:read</value>
+                                <value>sylius:shop:country:show</value>
                             </values>
                         </value>
                     </values>
@@ -111,7 +111,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>shop:country:read</value>
+                                <value>sylius:shop:country:show</value>
                             </values>
                         </value>
                     </values>
