+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/Currency.xml
@@ -32,7 +32,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:currency:read</value>
+                                <value>sylius:admin:currency:show</value>
                             </values>
                         </value>
                     </values>
@@ -44,7 +44,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:currency:read</value>
+                                <value>sylius:admin:currency:index</value>
                             </values>
                         </value>
                     </values>
@@ -56,7 +56,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:currency:create</value>
+                                <value>sylius:admin:currency:create</value>
                             </values>
                         </value>
                     </values>
@@ -65,7 +65,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>admin:currency:read</value>
+                                <value>sylius:admin:currency:show</value>
                             </values>
                         </value>
                     </values>
@@ -77,7 +77,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>shop:currency:read</value>
+                                <value>sylius:shop:currency:show</value>
                             </values>
                         </value>
                     </values>
@@ -89,7 +89,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>shop:currency:read</value>
+                                <value>sylius:shop:currency:show</value>
                             </values>
                         </value>
                     </values>
