+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/resources/ShopUserResetPassword.xml
@@ -44,7 +44,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>shop:reset_password:create</value>
+                                <value>sylius:shop:reset_password:create</value>
                             </values>
                         </value>
                     </values>
@@ -67,7 +67,7 @@
                     <values>
                         <value name="groups">
                             <values>
-                                <value>shop:reset_password:update</value>
+                                <value>sylius:shop:reset_password:update</value>
                             </values>
                         </value>
                     </values>
