+++ src/Sylius/Bundle/ApiBundle/Resources/config/services/serializers.xml
@@ -137,11 +137,11 @@
 <!--            <tag name="serializer.normalizer" priority="64" />-->
 <!--        </service>-->
 
-<!--        <service id="sylius.api.denormalizer.numeric_to_string.exchange_rate" class="Sylius\Bundle\ApiBundle\Serializer\NumericToStringDenormalizer">-->
-<!--            <argument type="string">%sylius.model.exchange_rate.class%</argument>-->
-<!--            <argument type="string">ratio</argument>-->
-<!--            <tag name="serializer.normalizer" priority="64" />-->
-<!--        </service>-->
+        <service id="sylius.api.denormalizer.numeric_to_string.exchange_rate" class="Sylius\Bundle\ApiBundle\Serializer\NumericToStringDenormalizer">
+            <argument type="string">%sylius.model.exchange_rate.class%</argument>
+            <argument type="string">ratio</argument>
+            <tag name="serializer.normalizer" priority="64" />
+        </service>
 
 <!--        <service id="Sylius\Bundle\ApiBundle\Serializer\CustomerDenormalizer">-->
 <!--            <argument type="service" id="clock" />-->
