+++ tests/Api/Admin/AdminUsersTest.php
@@ -13,12 +13,12 @@ declare(strict_types=1);
 
 namespace Sylius\Tests\Api\Admin;
 
+use Sylius\Component\Core\Model\AdminUserInterface;
 use Sylius\Tests\Api\JsonApiTestCase;
-use Sylius\Tests\Api\Utils\ContentType;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-final class AdminResetPasswordTest extends JsonApiTestCase
+final class AdminUsersTest extends JsonApiTestCase
 {
     /** @test */
     public function it_sends_administrator_password_reset_email(): void
@@ -27,33 +27,36 @@ final class AdminResetPasswordTest extends JsonApiTestCase
 
         $this->client->request(
             method: Request::METHOD_POST,
-            uri: '/api/v2/admin/reset-password',
+            uri: '/api/v2/admin/administrators/reset-password',
             server: self::CONTENT_TYPE_HEADER,
             content: json_encode([
                 'email' => 'api@example.com',
-            ], JSON_THROW_ON_ERROR),
+            ], \JSON_THROW_ON_ERROR),
         );
 
         $response = $this->client->getResponse();
-        self::assertEmailCount(1);
         $this->assertResponseCode($response, Response::HTTP_ACCEPTED);
     }
 
     /** @test */
-    public function it_resets_admin_user_password_with_valid_token(): void
+    public function it_resets_administrator_password(): void
     {
-        $this->loadFixturesFromFiles(['authentication/api_administrator.yaml']);
+        $loadedData = $this->loadFixturesFromFile('authentication/api_administrator.yaml');
 
-        $validToken = 'valid_token';
+        /** @var AdminUserInterface $adminUser */
+        $adminUser = $loadedData['admin'];
+        $adminUser->setPasswordResetToken('token');
+        $adminUser->setPasswordRequestedAt(new \DateTime('now'));
+        $this->getEntityManager()->flush();
 
         $this->client->request(
-            method: Request::METHOD_PATCH,
-            uri: sprintf('/api/v2/admin/reset-password/%s', $validToken),
-            server: ContentType::APPLICATION_JSON_MERGE_PATCH,
+            method: 'PATCH',
+            uri: '/api/v2/admin/administrators/reset-password/token',
+            server: self::PATCH_CONTENT_TYPE_HEADER,
             content: json_encode([
                 'newPassword' => 'newPassword',
                 'confirmNewPassword' => 'newPassword',
-            ], JSON_THROW_ON_ERROR),
+            ], \JSON_THROW_ON_ERROR),
         );
 
         $response = $this->client->getResponse();
