+++ tests/Api/Responses/admin/avatar_image/get_an_avatar_image.json
@@ -1,8 +1,8 @@
 {
-    "@context": "/api/v2/contexts/AvatarImage",
-    "@id": "/api/v2/admin/avatar-images/@integer@",
+    "@context": "\/api\/v2\/contexts\/AvatarImage",
+    "@id": "\/api\/v2\/admin\/administrators\/@integer@\/avatar-image",
     "@type": "AvatarImage",
     "id": @integer@,
     "path": "@string@.contains('sylius_original').endsWith('avatar_image.jpg')",
-    "owner": "\/api\/v2\/admin\/administrators/@integer@"
+    "owner": "\/api\/v2\/admin\/administrators\/@integer@"
 }
