+++ tests/Api/Responses/shop/currency/get_currencies_response.json
@@ -8,19 +8,7 @@
             "@type": "Currency",
             "code": "USD",
             "name": "US Dollar"
-        },
-        {
-            "@id": "\/api\/v2\/shop\/currencies\/GBP",
-            "@type": "Currency",
-            "code": "GBP",
-            "name": "British Pound"
-        },
-        {
-            "@id": "\/api\/v2\/shop\/currencies\/EUR",
-            "@type": "Currency",
-            "code": "EUR",
-            "name": "Euro"
         }
     ],
-    "hydra:totalItems": 3
+    "hydra:totalItems": 1
 }
