+++ tests/Api/Shop/CurrenciesTest.php
@@ -22,10 +22,10 @@ final class CurrenciesTest extends JsonApiTestCase
     /** @test */
     public function it_gets_a_currency(): void
     {
-        $fixtures = $this->loadFixturesFromFiles(['currency.yaml']);
+        $fixtures = $this->loadFixturesFromFiles(['channel.yaml', 'currency.yaml']);
 
         /** @var CurrencyInterface $currency */
-        $currency = $fixtures['currency_gbp'];
+        $currency = $fixtures['currency_usd'];
 
         $this->client->request(
             method: 'GET',
